BAH-4659|Fix. move inventory services into bahmni-services
