+++ apps/clinical/src/providers/clinicalConfig/models.ts
@@ -8,7 +8,6 @@ export interface AllergyConceptMap {
 export interface InputControlAttributes {
   name: string;
   required: boolean;
-  config?: Record<string, unknown>;
 }
 
 export interface InputControl<
