+++ packages/bahmni-services/src/inventoryService/index.ts
@@ -1,0 +1,3 @@
+export { getAvailableStocks } from './inventoryService';
+export type { StockBatch, AvailableStockResponse } from './models';
+export { AVAILABLE_STOCKS_URL } from './constants';
