+++ packages/bahmni-services/src/inventoryService/models.ts
@@ -1,0 +1,13 @@
+export interface StockBatch {
+  stockLocationName: string;
+  availableQuantity: number;
+  onHandQuantity: number;
+  unit: string;
+  batchNumber: string;
+  expiryDate: string;
+}
+
+export interface AvailableStockResponse {
+  count: number;
+  data: StockBatch[];
+}
